Move and fix file paths for App Component files.
