+++ src/Components/App/App.js
@@ -1,0 +1,20 @@
+import './App.css';
+import {Component} from 'react'
+import {BrowserRouter} from 'react-router-dom'
+class App extends Component {
+  constructor(props){
+    super(props)
+    this.state= {
+
+    }
+  }
+  render () {
+    return (
+      <div className="App">
+        <h1>Test</h1>
+      </div>
+    );
+  }
+}
+
+export default App;
